requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.1
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.5
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.1.7
django-crispy-forms==1.11.2
django-filter==2.4.0
django-weasyprint==1.1.0
future==0.18.2
html5lib==1.1
Pillow==8.1.2
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.10.0
python-bidi==0.4.2
pytz==2021.1
reportlab==3.5.66
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
WeasyPrint==52.4
webencodings==0.5.1
